intelli/requirements.txt
# Core dependencies
python-dotenv~=1.0.1
numpy<2.0
mcp~=1.9.0
matplotlib>=3.6.0
networkx>=3.2.1
huggingface_hub>=0.28.1

# ML dependencies
keras-nlp
keras>=3
librosa
keras-hub
torch>=2.6.0
safetensors>=0.5.0
triton>=2.1.0

# For testing
pytest>=7.0.0

# Other dependencies
asyncio
json5